Addressing complexity around cli option parsing
